Header component, navigation layout and themes menu.
