+++ src/components/layout/Layout.tsx
@@ -1,22 +1,22 @@
 import { PreviewAlert } from 'src/components/PreviewAlert'
 import { Container } from 'hds-react'
 import styles from './layout.module.scss'
-import Nav from 'src/components/Nav'
-import { DrupalMenuLinkContent } from "next-drupal"
+import Header from "src/components/Header"
+import { HeaderProps } from "src/lib/types"
 
 interface LayoutProps {
   children: any,
-  menu: DrupalMenuLinkContent[],
+  header: HeaderProps,
 }
 
-export function Layout({ children, menu }: LayoutProps): JSX.Element {
+export function Layout({ children, header }: LayoutProps): JSX.Element {
   return (
     <>
       <PreviewAlert />
       <div className={styles.wrapper}>
         <header>
           <div>
-          <Nav menu={menu}/>
+          <Header {...header} />
           </div>
         </header>
         <main>{children}</main>
